Add component_image
